requirements.txt
paho-mqtt
lgpio
pirlib
